统一md5计算
