+++ config/src/main/java/com/alibaba/nacos/config/server/service/dump/DumpService.java
@@ -16,6 +16,7 @@
 package com.alibaba.nacos.config.server.service.dump;
 
 import com.alibaba.nacos.common.utils.IoUtils;
+import com.alibaba.nacos.common.utils.Md5Utils;
 import com.alibaba.nacos.config.server.constant.Constants;
 import com.alibaba.nacos.config.server.manager.TaskManager;
 import com.alibaba.nacos.config.server.model.ConfigInfo;
@@ -346,7 +347,8 @@ public class DumpService {
                         ConfigInfo cf = MergeTaskProcessor.merge(dataId, group, tenant, datumList);
                         String aggrContent = cf.getContent();
                         String localContentMD5 = ConfigService.getContentMd5(GroupKey.getKey(dataId, group));
-                        String aggrConetentMD5 = MD5.getInstance().getMD5String(aggrContent);
+                        String aggrConetentMD5 = Md5Utils.getMD5(aggrContent, Constants.ENCODE);
+
                         if (!StringUtils.equals(localContentMD5, aggrConetentMD5)) {
                             persistService.insertOrUpdate(null, null, cf, time, null, false);
                             log.info("[merge-ok] {}, {}, size={}, length={}, md5={}, content={}", dataId, group,
